+++ src/Stack/Types/Build.hs
@@ -96,6 +96,7 @@ data StackBuildException
         (Maybe (CompilerVersion, Arch))
         (CompilerVersion, Arch)
         GHCVariant
+        CompilerBuild
         VersionCheck
         (Maybe (Path Abs File))
         Text -- recommended resolution
@@ -145,7 +146,7 @@ instance Show StackBuildException where
                ", the package id couldn't be found " <> "(via ghc-pkg describe " <>
                packageNameString name <> "). This shouldn't happen, " <>
                "please report as a bug")
-    show (CompilerVersionMismatch mactual (expected, earch) ghcVariant check mstack resolution) = concat
+    show (CompilerVersionMismatch mactual (expected, earch) ghcVariant ghcBuild check mstack resolution) = concat
                 [ case mactual of
                     Nothing -> "No compiler found, expected "
                     Just (actual, arch) -> concat
@@ -164,6 +165,7 @@ instance Show StackBuildException where
                 , " ("
                 , C.display earch
                 , ghcVariantSuffix ghcVariant
+                , compilerBuildSuffix ghcBuild
                 , ") (based on "
                 , case mstack of
                     Nothing -> "command line arguments"